Dev -> Main

+++ app/src/main/java/com/romandevyatov/bestfinance/viewmodels/foreachmodel/IncomeSubGroupViewModel.kt
@@ -3,8 +3,8 @@ package com.romandevyatov.bestfinance.viewmodels.foreachmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.romandevyatov.bestfinance.db.entities.IncomeSubGroup
-import com.romandevyatov.bestfinance.repositories.IncomeSubGroupRepository
+import com.romandevyatov.bestfinance.data.entities.IncomeSubGroup
+import com.romandevyatov.bestfinance.data.repositories.IncomeSubGroupRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -30,7 +30,7 @@ class IncomeSubGroupViewModel @Inject constructor(
         incomeSubGroupRepository.deleteIncomeSubGroup(incomeSubGroup)
     }
 
-    fun deleteIncomeSubGroupById(id: Int) = viewModelScope.launch(Dispatchers.IO) {
+    fun deleteIncomeSubGroupById(id: Long) = viewModelScope.launch(Dispatchers.IO) {
         incomeSubGroupRepository.deleteIncomeSubGroupById(id)
     }
